heart anomly detection Model training/heart anomly detection Model training /arrhythmia_risk_model/requirements.txt
# Guardian Angel - Arrhythmia Risk Screening Model
# Python 3.10+ | macOS Apple Silicon Compatible
# Last updated: 2026-01-03

# =============================================================================
# CORE DATA PROCESSING
# =============================================================================
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
scipy>=1.11.0,<2.0.0

# =============================================================================
# ECG/PHYSIOLOGICAL DATA
# =============================================================================
wfdb>=4.1.0,<5.0.0              # MIT-BIH database access & ECG processing

# =============================================================================
# MACHINE LEARNING
# =============================================================================
scikit-learn>=1.3.0,<2.0.0      # Preprocessing, metrics, cross-validation
xgboost>=2.0.0,<3.0.0           # Primary model (CPU-only, no GPU dependency)

# =============================================================================
# HRV ANALYSIS (OPTIONAL BUT RECOMMENDED)
# =============================================================================
antropy>=0.1.6                   # Entropy measures for HRV non-linear analysis

# =============================================================================
# CONFIGURATION & UTILITIES
# =============================================================================
pyyaml>=6.0.0,<7.0.0            # Configuration file parsing
python-dotenv>=1.0.0,<2.0.0     # Environment variable management

# =============================================================================
# MODEL SERIALIZATION & EXPORT
# =============================================================================
joblib>=1.3.0,<2.0.0            # Model persistence
pyarrow>=14.0.0,<16.0.0         # Parquet support for feature storage

# =============================================================================
# EVALUATION & VISUALIZATION (OPTIONAL)
# =============================================================================
matplotlib>=3.7.0,<4.0.0        # Evaluation plots
seaborn>=0.13.0,<1.0.0          # Statistical visualization

# =============================================================================
# DEVELOPMENT & TESTING
# =============================================================================
pytest>=7.4.0,<8.0.0            # Unit testing framework
pytest-cov>=4.1.0,<5.0.0        # Coverage reporting

# =============================================================================
# LOGGING & EXPERIMENT TRACKING
# =============================================================================
loguru>=0.7.0,<1.0.0            # Structured logging

# =============================================================================
# TYPE CHECKING (OPTIONAL)
# =============================================================================
# mypy>=1.5.0                   # Static type checking (uncomment if needed)
